Automated budgets working
I've added an automated income/expense budgeting system that will simulate the
effect of these transactions on your accounts. The interface, however, is
pretty atrocious; hopefully I can make that a bit better. The calculations on
the applied budgets' running balance are also a bit off, but this appears to
be out-of-order transactions, something I can't immediately solve (as Postings
have no awareness of order). For now, they'll start at 0.