Posting delete test, removed Posting edit stuff
In addition to testing the Posting delete cascade, I also removed a significant
chunk of posting_post_save and all of posting_pre_save. The code to change a
Posting's balance and the corresponding account balance was not behaving as
expected. In order to provide a better audit trail and show the edits being
made, I'll have to create a new editing mechanism that preserves the old data.
This will also apply to deletion in the future, as that was mainly for testing
purposes. Eventually, all the changes made on Postings will be visible on
request.