Modified Posting signal handlers, budget test fail
I'm in the middle of reworking the Posting signal handlers to properly delete
the opposite Posting and Journal. This creates recursion depth problems in
places, as the delete cascade keeps triggering the signals, which triggers the
delete, and so on. I've reached a point where the accounts tests are passing,
which means the Posting and its complementary negative are both deleted. I may
have solved the recursion problem by temporarily disconnecting the signal
handler; hopefully this doesn't cause problems when I do need a signal
triggered. However, the Budget is_applied_switch test is failing hard, so I would
consider Budgets unreliable for the time being. Next on the agenda is to work out
the Journal deletion process, without triggering Journal.DoesNotExist.